+++ apps/admin/src/features/product/components/FormActions.module.scss
@@ -1,0 +1,9 @@
+.buttons {
+  @include flex-center;
+  gap: 12px;
+  position: sticky;
+  bottom: 0;
+  padding: 24px;
+  background-color: #fff;
+  border-top: 1px solid #f5f5f5;
+}
