+++ apps/admin/src/features/product/components/ProductBasicForm.module.scss
@@ -1,0 +1,5 @@
+.container {
+  @include flex-col;
+  gap: 10px;
+  padding: 24px;
+}
