+++ apps/admin/src/features/product/components/ProductImageUploader.module.scss
@@ -1,0 +1,5 @@
+.description {
+  font-size: $font-xs;
+  color: $gray-dark;
+  line-height: 1.5;
+}
