+++ apps/admin/src/features/product/components/index.ts
@@ -1,1 +1,5 @@
+export * from './FormActions';
 export * from './OptionSizeTable';
+export * from './ProductImageUploader';
+export * from './ProductOptionForm';
+export * from './ProductBasicForm';
