+++ apps/admin/src/features/product/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './queries';
 
 export * from './useOptionSize';
 export * from './useImageUploader';
+export * from './useProductRegisterForm';
