+++ apps/admin/src/pages/ProductRegister.module.scss
@@ -4,50 +4,3 @@
   height: 100vh;
   overflow: auto;
 }
-
-.container {
-  @include flex-col;
-  gap: 10px;
-  padding: 24px;
-}
-
-.buttons {
-  @include flex-center;
-  gap: 12px;
-  position: sticky;
-  bottom: 0;
-  padding: 24px;
-  background-color: #fff;
-  border-top: 1px solid #f5f5f5;
-}
-
-.size__buttons {
-  @include flex-item;
-  flex-wrap: wrap;
-  width: 100%;
-  gap: 12px;
-}
-
-.size {
-  @include flex-col;
-  gap: 24px;
-}
-
-.description {
-  font-size: $font-xs;
-  color: $gray-dark;
-  line-height: 1.5;
-}
-
-.size__title {
-  font-size: $font-sm;
-  font-weight: 600;
-  padding-bottom: 12px;
-}
-
-.stock__input {
-  @include flex-item;
-  justify-content: flex-end;
-  gap: 24px;
-  width: 100%;
-}
